some easy questions solved, js 30 folder added

+++ easy/169.py
@@ -1,4 +1,4 @@
 class Solution:
     def majorityElement(self, nums: list[int]) -> int:
         nums.sort()
-        return nums[len(nums)/2]
+        return nums[int(len(nums)/2)]
